Remove "refresh" thread and just set S3 user script ACL to public-read
Fixes #1600, and also fixes #1516. As a side effect, the expiration
isn't really enforced. NB: the extra fiddling to remove the URL
signing seems to be necessary: even a publicly-available URL will
return a 403 if the signing token has expired.

This also removes a few unused imports in jobStores/aws/jobStore.py.