Avoid checking deadlocks when processing many finished leaf jobs
Jobs that don't update their parent or launch new jobs caused a
deadlock check. These deadlock checks are actually very expensive now,
because they check for the existence of flag files in the jobStore.

The intention was clearly that deadlock checks would only occur when
"nothing happened" in any particular run of the loop, but the check
was incorrect: it checked for toilState.updatedJobs to be empty, but
that only indicates that no *new* jobs were updated (after potentially
processing a finished job). In the case that many leaf jobs come in
from the batch system at once, a deadlock check will be incurred
after *each one*, causing huge slowdowns if you have services running.