Config Grafana to access Prom via loopback instead of via user browser
This should make it a bit easier to view multiple metrics dashboards
if you are running multiple workflows at once. It also removes the
need to forward port 9090. Previously, accessing a Grafana instance,
even on a different port, would attempt to access Prometheus *from the
user's browser*--meaning that if you are running workflows A and B,
and A's Prometheus instance was forwarded to localhost:9090, B's
grafana dashboard would always show the data from *workflow A*.